Fix integer-overflow bypass of the map render dimension clamp (audit HIGH-1)
Chunk coordinates come from attacker-controlled region filenames (r.<X>.<Z>.mca
in a pushed manifest) and span the full int range. The 160-chunk clamp tested
`maxCX - minCX + 1 > MaxSideChunks` in int: for a span near 2^31 the subtraction
wrapped negative, the test was false, the clamp was skipped, and the wrapped
span flowed into `new byte[w*h*3]` — a multi-hundred-MB-to-negative allocation
(OOM, or OverflowException → 500). Any user who can push could trigger it by
rendering a world with two region files at extreme coords (2 chunks, under the
chunk cap).

Compute the span and recenter midpoint in long, and clamp before narrowing to
int; after clamping each span is provably ≤ MaxSideChunks, so w,h ≤ 2560 and the
int math can't overflow. Regression test renders a world with chunks ~4 billion
apart and asserts the canvas stays bounded instead of crashing.

+++ src/McadiffHub/MapRenderer.cs
@@ -51,9 +51,13 @@ public static class MapRenderer
         if (surfaces.Count == 0) { info = new MapInfo(0, 0, 0, false, regionsRead); return EmptyPng(); }
 
         // Clamp a runaway span to a window centered on the populated area. A hit chunk-cap also truncates.
+        // Chunk coords come from attacker-controlled region filenames (r.<X>.<Z>.mca) and span the full int
+        // range, so the span and midpoint are computed in LONG: `max - min + 1` in int would overflow and
+        // wrap past the `> MaxSideChunks` test, bypassing the clamp into a huge/negative allocation. After
+        // clamping, each span is provably ≤ MaxSideChunks, so the int math below can't overflow (w,h ≤ 2560).
         bool truncated = capHit;
-        if (maxCX - minCX + 1 > MaxSideChunks) { int c = (minCX + maxCX) / 2; minCX = c - MaxSideChunks / 2; maxCX = minCX + MaxSideChunks - 1; truncated = true; }
-        if (maxCZ - minCZ + 1 > MaxSideChunks) { int c = (minCZ + maxCZ) / 2; minCZ = c - MaxSideChunks / 2; maxCZ = minCZ + MaxSideChunks - 1; truncated = true; }
+        if ((long)maxCX - minCX + 1 > MaxSideChunks) { long c = ((long)minCX + maxCX) / 2; minCX = (int)(c - MaxSideChunks / 2); maxCX = minCX + MaxSideChunks - 1; truncated = true; }
+        if ((long)maxCZ - minCZ + 1 > MaxSideChunks) { long c = ((long)minCZ + maxCZ) / 2; minCZ = (int)(c - MaxSideChunks / 2); maxCZ = minCZ + MaxSideChunks - 1; truncated = true; }
 
         int w = (maxCX - minCX + 1) * 16, h = (maxCZ - minCZ + 1) * 16;
         byte[] rgb = new byte[w * h * 3];
